Split component collection classes out of EntityManager.
